Fold Bootstrap adapter resolution into Inertia.SSR
Bootstrap was @moduledoc false, had a single caller (Inertia.SSR.init/1),
and existed mostly as a holdover from earlier iterations. Inline its
adapter resolution as private helpers on Inertia.SSR and drop the file.

The three adapter-validation tests now call Inertia.SSR.init/1
directly — the real code path the supervisor takes — instead of the
intermediate Bootstrap.fetch_adapter/2. init/1 raises before any
:persistent_term.put or Supervisor.init runs, so failure-case tests
stay side-effect-free.
